change to float64

+++ ssnsp/helper/data_generation.py
@@ -35,9 +35,9 @@ def lasso_test(N = 10, n = 20, k = 5, lambda1 = .1, block = False):
     # create measurements
     b = A @ x
     
-    A = A.astype('float32')
-    b = b.astype('float32')
-    x = x.astype('float32')
+    A = A.astype('float64')
+    b = b.astype('float64')
+    x = x.astype('float64')
     
     phi = Norm1(lambda1)    
     if block:
@@ -70,10 +70,6 @@ def logreg_test(N = 10, n = 20, k = 5, lambda1 = .1):
     odds = h/(1+h)
     
     b = (odds >= .5)*2 -1
-    
-    A = A.astype('float32')
-    b = b.astype('float32')
-    x = x.astype('float32')
     
     phi = Norm1(lambda1) 
     f = logistic_loss(A,b)
